Amend OV5647 processing to not include ver char in search
In the Broadcom header, the search string should be
BRCM, and it is then followed by a header version character.
That character can change as the firmware is updated,
although the structure is generally kept as backwards
compatible as possible.
Remove the version char from the search string in identify()